refactor datalog emit code utility. fix things that don't have tests that are affected by refactoring
