+++ raco/from_repr.py
@@ -2,7 +2,7 @@
 
 from raco.backends.myria import *
 from raco.backends.cpp import *
-from raco.backends.cpp.clangcommon import *
+from raco.backends.cpp.cppcommon import *
 from raco.backends.radish import *
 from raco.algebra import *
 from raco.scheme import *
